. blkalloc.cpp, blkalloc.h, regfs.cpp: fixed a bug in IntersectIntervals, which caused merging blocks to fail sometimes when the block to write intersecting the last block in the dirty map, And fixed a bug in SubtractIntervals, which set wrong offset into the buffer to read/write. . regfschk.cpp: fixed a bug in the file size check, which report false alert when the file size equals MAX_FILE_SIZE, and caused the file are treated as corrupted file. . monfuse.cpp, blkalloc.h: explicitly set MAX_FS_SIZE to 64bit since it can reach 128GB as the MAX_FILE_SIZE.

+++ monitor/registry/blkalloc.h
@@ -217,7 +217,7 @@
 #define GID_DEFAULT     80001
 
 #define MAX_FS_SIZE     \
-    ( ( ( BLOCKS_PER_GROUP_FULL * BLKGRP_NUMBER_FULL ) + 1 + 1 ) * BLOCK_SIZE )
+    ( ( ( ( guint64 )BLOCKS_PER_GROUP_FULL * BLKGRP_NUMBER_FULL ) + 1 + 1 ) * BLOCK_SIZE )
 
 #define BLKIDX_BEYOND_MAX( _blkidx_  ) \
     ( ( _blkidx_ ) >= BLKGRP_NUMBER_FULL * BLOCKS_PER_GROUP_FULL )
